+++ lib/db_mod/statements/configuration/single/required_value.rb
@@ -14,7 +14,7 @@ module DbMod
         #  end
         module RequiredValue
           # Enables this module to be passed to
-          # {DbMod::Statements::Configuration.process_method_results} as the
+          # {DbMod::Statements::Configuration.attach_result_processor} as the
           # +wrapper+ function, where it will return the first column of the
           # first row of the result set, or fail if anything other than
           # exactly one row is returned.